pyproject.toml
[tool.poetry]
name = "pandasai-chainlit-docker-template"
version = "0.1.0"
description = "A template PandasAI Chainlit App with Docker Deployments"
authors = ["Amjad Raza"]
license = "MIT"
readme = "README.md"
packages = [{include = "pandasai_chainlit_docker_deployment_template"}]


[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
chainlit = "^0.3.0"
pandasai = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
